move non shared types from shared to where they belong. Also improved some comments.
