+++ packages/frontend/src/components/helpers/MapMsgStatus.ts
@@ -1,5 +1,5 @@
 import { C } from '@deltachat/jsonrpc-client'
-import { msgStatus } from '../../../../shared/shared-types'
+import { msgStatus } from '../../types-app'
 
 export function mapCoreMsgStatus2String(state: number): msgStatus {
   switch (state) {
