+++ packages/frontend/src/types-app.d.ts
@@ -1,0 +1,33 @@
+// moved from shared/shared-types because only used by frontend.
+// TODO: find a better place than this file for those types
+
+export type Credentials = {
+  [key: string]: any
+  addr?: string
+  mail_user?: string
+  mail_pw?: string
+  mail_server?: string
+  mail_port?: string
+  mail_security?: 'automatic' | '' | 'ssl' | 'default'
+  imap_certificate_checks?: any
+  send_user?: string
+  send_pw?: string
+  send_server?: string
+  send_port?: string
+  send_security?: 'automatic' | '' | 'ssl' | 'starttls' | 'plain'
+  smtp_certificate_checks?: any
+  socks5_enabled: '1' | '0'
+  socks5_host: string
+  socks5_port: string
+  socks5_user: string
+  socks5_password: string
+}
+
+export type msgStatus =
+  | 'error'
+  | 'sending'
+  | 'draft'
+  | 'delivered'
+  | 'read'
+  | 'sent'
+  | ''
